refactor(sandbox): split process.rs into codec/handshake/spawn/dispatch per §12.7
Mechanical, behavior-preserving split of the ~2209-line process.rs god-file
into four focused modules plus moving the SandboxError→ActionError converter
into error.rs. No public-API change: `pub use process::ProcessSandbox` becomes
`pub use dispatch::ProcessSandbox` with byte-identical external names.

- codec.rs: bounded line framing, PluginHandle, stderr drain (no spawn knowledge)
- handshake.rs: socket allocation, addr validation, correlation-id extraction
- spawn.rs: the only module with `unsafe` (the pre_exec hardening block)
- dispatch.rs: ProcessSandbox, TryDispatchError, race_cancel_timeout, trait impls
- error.rs: sandbox_error_to_action_error moved here as pub(crate)

spawn_and_dial converted from a &self method to a free pub(crate) fn taking
(binary, linux_rlimits). Cross-module seams widened to pub(crate). Every test
preserved (57 passed, 1 skipped — unchanged from baseline).
