--- crates/sandbox/src/process.rs
@@ -1,2209 +1,0 @@
-//! Process-based sandbox for community plugins using the duplex v3 protocol.
-//!
-//! Slice 1c (2026-04-13): plugin processes are **long-lived**. On the first
-//! call, `ProcessSandbox` spawns the plugin binary, reads the handshake line
-//! from its stdout, dials the announced UDS or Named Pipe, and stores the
-//! resulting [`PluginHandle`] on the sandbox. Subsequent calls reuse that
-//! handle, sending envelopes over the socket without respawning. A broken
-//! connection (plugin crashed or exited) clears the handle and the next
-//! request triggers a fresh spawn.
-//!
-//! The plugin-side event loop in `nebula-plugin-sdk::run_duplex` is still
-//! sequential — one action at a time per plugin process. Slice 1d adds
-//! concurrent multiplexed dispatch.
-//!
-//! Security enforcement (unchanged since slice 1b):
-//! - `env_clear()` + explicit env allowlist
-//! - `pre_exec` landlock + configurable rlimits (Linux)
-//! - stderr size limit for log capture
-//! - `kill_on_drop` on the spawned child → plugin process dies with the sandbox
-//!
-//! ## Transport line-length caps (#316, 2026-04-14)
-//!
-//! Both the handshake and envelope read paths are length-capped. An
-//! untrusted plugin that emits a newline-starved or gigabyte-sized line no
-//! longer grows the receive buffer until OOM: the read returns a typed
-//! [`SandboxError::PluginLineTooLarge`] / [`SandboxError::HandshakeLineTooLarge`]
-//! and the plugin transport is **poisoned** — further send/recv calls on
-//! the same handle fail fast with [`SandboxError::TransportPoisoned`]. The
-//! outer `dispatch_envelope` also drops its cached handle on any error,
-//! so defense-in-depth gives us both an instance-level invalidation
-//! (`poisoned` flag inside `PluginHandle`) and a sandbox-level one
-//! (`*self.handle.lock().await = None`).
-
-use std::{
-    path::PathBuf,
-    sync::atomic::{AtomicU64, Ordering},
-    time::Duration,
-};
-
-use async_trait::async_trait;
-use nebula_action::{ActionError, ActionMetadata, result::ActionResult};
-use nebula_plugin_sdk::{
-    protocol::{HostToPlugin, PluginToHost},
-    transport::{self, ENV_SOCKET_ADDR, ENV_SOCKET_KIND, PluginStream},
-};
-use tokio::{
-    io::{AsyncBufRead, AsyncBufReadExt, AsyncReadExt, AsyncWriteExt, BufReader},
-    process::{Child, Command},
-    sync::Mutex,
-};
-use tokio_util::sync::CancellationToken;
-
-use crate::{
-    SandboxRunner, error::SandboxError, os_sandbox::LinuxRlimits, runner::SandboxedContext,
-};
-
-/// Timeout for reading the plugin's handshake line from stdout.
-const HANDSHAKE_TIMEOUT: Duration = Duration::from_secs(3);
-
-/// Maximum bytes accepted for the plugin handshake line.
-///
-/// A handshake is a short socket/pipe address string plus protocol version
-/// — realistically under 100 bytes. 4 KiB gives ~40x headroom while still
-/// bounding memory use far below anything that could stress the allocator.
-const HANDSHAKE_LINE_CAP: usize = 4 * 1024;
-
-/// Maximum bytes accepted for a single runtime envelope line
-/// (JSON payload + trailing `\n`).
-///
-/// 1 MiB starting point. If real plugins need more, make this configurable
-/// via [`ProcessSandbox::new`] later rather than preemptively raising the
-/// ceiling — a too-high cap defeats the purpose of the cap.
-const ENVELOPE_LINE_CAP: usize = 1024 * 1024;
-
-/// Maximum bytes buffered per plugin stderr log line before the overflow
-/// is silently discarded (stderr is diagnostic log output, not protocol,
-/// so truncation is acceptable here — but we still bound the read, because
-/// an attacker-controlled plugin could otherwise emit a gigabyte of
-/// newline-starved garbage and starve the host of memory).
-const STDERR_LINE_CAP: usize = 8 * 1024;
-
-/// Process sandbox: spawns the plugin binary once and keeps the connection
-/// alive for the lifetime of this sandbox instance.
-///
-/// Each `ProcessSandbox` owns a long-lived `PluginHandle` behind a
-/// `Mutex`. The first invocation spawns the child and dials the socket;
-/// subsequent invocations reuse the same handle. A connection error on
-/// write or read invalidates the handle and the next call respawns.
-pub struct ProcessSandbox {
-    /// Path to the plugin binary.
-    binary: PathBuf,
-    /// Per-call timeout (envelope round-trip wall clock).
-    timeout: Duration,
-    /// Linux child-process resource limits (ignored on non-Linux).
-    linux_rlimits: LinuxRlimits,
-    /// Long-lived handle to the spawned plugin process. Serialized via the
-    /// mutex — slice 1c is sequential per sandbox instance. Slice 1d can
-    /// replace this with a lock-free handle once concurrent dispatch lands.
-    handle: Mutex<Option<PluginHandle>>,
-    /// Monotonic correlation id source (#285). Each outbound envelope
-    /// gets a fresh id; `try_dispatch` verifies the response echoes it
-    /// back. A stale response (e.g. late reply to a cancelled call)
-    /// therefore can't be silently mis-associated with a fresh request
-    /// — ID mismatch poisons the transport.
-    ///
-    /// Persisted across plugin respawns — the plugin only sees a
-    /// monotone sequence from its own perspective (fresh process,
-    /// fresh socket, fresh id stream), but the host never reuses an
-    /// id across invocations within this sandbox instance's lifetime.
-    next_id: AtomicU64,
-}
-
-/// Live connection to a running plugin process.
-///
-/// Owns the spawned [`Child`] and the two halves of the accepted stream
-/// (reader is wrapped in `BufReader` for efficient line-delimited reads).
-/// When dropped, `kill_on_drop(true)` on the child ensures the OS process
-/// is terminated; the socket/pipe is released by `PluginStream`'s cleanup
-/// guard on the plugin side.
-struct PluginHandle {
-    /// Kept alive for `kill_on_drop` — dropping this struct SIGKILLs the
-    /// child. Read nowhere; the underscore prefix silences dead-code
-    /// warnings.
-    _child: Child,
-    /// Host-allocated temp directory holding the UDS socket on Unix.
-    /// Kept alive for the lifetime of the handle so the directory (and
-    /// its 0700 perms) survive as long as the plugin process. Dropping
-    /// the `TempDir` removes the directory tree. `None` on Windows —
-    /// named pipes aren't in a filesystem directory. Dead code at
-    /// read-time; the `_` prefix silences warnings.
-    _socket_dir: Option<tempfile::TempDir>,
-    /// Buffered reader over the stream's read half. Crucial for
-    /// throughput — byte-at-a-time reads hit ~4 MB/s, BufReader reaches
-    /// hundreds of MB/s on local sockets/pipes.
-    reader: BufReader<tokio::io::ReadHalf<PluginStream>>,
-    /// Owning write half for envelope dispatch.
-    writer: tokio::io::WriteHalf<PluginStream>,
-    /// Scratch byte buffer reused across `recv_envelope` calls to avoid
-    /// per-call allocation. `Vec<u8>` rather than `String` so `read_until`
-    /// can drive the bounded read without a UTF-8 validation in the hot
-    /// path — parsing happens via `serde_json::from_slice` after the cap
-    /// check.
-    line_buf: Vec<u8>,
-    /// Set to `true` once the transport has produced any error that
-    /// leaves us at an unknown position in the byte stream (oversized
-    /// line, mid-frame I/O failure). Once poisoned, every subsequent
-    /// `send_envelope` / `recv_envelope` call short-circuits with
-    /// [`SandboxError::TransportPoisoned`] instead of reading more bytes.
-    /// This is defense-in-depth alongside the outer
-    /// `ProcessSandbox::try_dispatch` handle-clear logic.
-    poisoned: bool,
-}
-
-impl PluginHandle {
-    fn new(child: Child, stream: PluginStream, socket_dir: Option<tempfile::TempDir>) -> Self {
-        let (read_half, write_half) = tokio::io::split(stream);
-        Self {
-            _child: child,
-            _socket_dir: socket_dir,
-            reader: BufReader::new(read_half),
-            writer: write_half,
-            line_buf: Vec::with_capacity(512),
-            poisoned: false,
-        }
-    }
-
-    async fn send_envelope(&mut self, envelope: &HostToPlugin) -> Result<(), SandboxError> {
-        if self.poisoned {
-            return Err(SandboxError::TransportPoisoned);
-        }
-        let encoded = serde_json::to_vec(envelope).map_err(SandboxError::HostMalformedEnvelope)?;
-        if let Err(e) = self.writer.write_all(&encoded).await {
-            self.poisoned = true;
-            return Err(SandboxError::Transport(e));
-        }
-        if let Err(e) = self.writer.write_all(b"\n").await {
-            self.poisoned = true;
-            return Err(SandboxError::Transport(e));
-        }
-        if let Err(e) = self.writer.flush().await {
-            self.poisoned = true;
-            return Err(SandboxError::Transport(e));
-        }
-        Ok(())
-    }
-
-    async fn recv_envelope(&mut self) -> Result<PluginToHost, SandboxError> {
-        if self.poisoned {
-            return Err(SandboxError::TransportPoisoned);
-        }
-        self.line_buf.clear();
-        match read_bounded_line(&mut self.reader, ENVELOPE_LINE_CAP, &mut self.line_buf).await {
-            Ok(BoundedReadOutcome::Line { body_len }) => {
-                let body = &self.line_buf[..body_len];
-                serde_json::from_slice::<PluginToHost>(body).map_err(|e| {
-                    // Parse error does not leave the stream in an unknown
-                    // state — we consumed exactly one line. Do NOT poison
-                    // here; the outer dispatch will still drop the handle
-                    // to be safe, but a legitimate one-off parse failure
-                    // should not prevent a retry on a fresh envelope.
-                    SandboxError::MalformedEnvelope(e)
-                })
-            },
-            Ok(BoundedReadOutcome::Eof) => {
-                self.poisoned = true;
-                Err(SandboxError::PluginClosed)
-            },
-            Ok(BoundedReadOutcome::Overflow { observed }) => {
-                self.poisoned = true;
-                Err(SandboxError::PluginLineTooLarge {
-                    limit: ENVELOPE_LINE_CAP,
-                    observed,
-                })
-            },
-            Err(e) => {
-                self.poisoned = true;
-                Err(SandboxError::Transport(e))
-            },
-        }
-    }
-
-    /// Receive the next envelope as raw JSON bytes (no typed parse).
-    ///
-    /// Used by the metadata-probe path (see `ProcessSandbox::get_metadata_raw`):
-    /// returning the raw bytes lets the caller first parse to
-    /// `serde_json::Value` for a cheap version check, then re-parse the
-    /// same bytes as `PluginToHost` — preserving the zero-copy / borrowed-
-    /// string path that some `Deserialize` implementations (notably
-    /// `domain_key::Key<T>`) require.
-    ///
-    /// Contract mirrors `recv_envelope`: same poisoning rules, same cap
-    /// enforcement, same error taxonomy; only the final parse step is
-    /// deferred to the caller.
-    async fn recv_envelope_bytes(&mut self) -> Result<Vec<u8>, SandboxError> {
-        if self.poisoned {
-            return Err(SandboxError::TransportPoisoned);
-        }
-        self.line_buf.clear();
-        match read_bounded_line(&mut self.reader, ENVELOPE_LINE_CAP, &mut self.line_buf).await {
-            Ok(BoundedReadOutcome::Line { body_len }) => Ok(self.line_buf[..body_len].to_vec()),
-            Ok(BoundedReadOutcome::Eof) => {
-                self.poisoned = true;
-                Err(SandboxError::PluginClosed)
-            },
-            Ok(BoundedReadOutcome::Overflow { observed }) => {
-                self.poisoned = true;
-                Err(SandboxError::PluginLineTooLarge {
-                    limit: ENVELOPE_LINE_CAP,
-                    observed,
-                })
-            },
-            Err(e) => {
-                self.poisoned = true;
-                Err(SandboxError::Transport(e))
-            },
-        }
-    }
-}
-
-/// Outcome of a single bounded line read.
-#[derive(Debug, PartialEq, Eq)]
-enum BoundedReadOutcome {
-    /// A complete newline-terminated line was read. `body_len` is the
-    /// length of the line **excluding** the trailing `\n`; callers slice
-    /// `&buf[..body_len]` directly, so the no-underflow / in-bounds
-    /// invariant lives in the type rather than in a recomputed `- 1` at
-    /// every call site.
-    Line { body_len: usize },
-    /// EOF reached before any bytes were read. The stream is closed.
-    Eof,
-    /// More than `cap` bytes were consumed without encountering a
-    /// newline. The stream position is now `cap + 1` bytes past where
-    /// the read began; the transport **must not be reused** because we
-    /// cannot resync on a line boundary without reading an unbounded
-    /// amount more.
-    Overflow {
-        /// Number of bytes actually consumed from the stream before the
-        /// cap was hit. Always `> cap`.
-        observed: usize,
-    },
-}
-
-/// Read a single newline-delimited line from `reader`, refusing to buffer
-/// more than `cap` bytes.
-///
-/// Uses `AsyncBufReadExt::take(cap + 1).read_until(b'\n', buf)` as the
-/// primitive. The `+ 1` is load-bearing: it lets us distinguish "exactly
-/// `cap` bytes including the newline, legal" from "more than `cap` bytes,
-/// cap breached". Without it we could not tell whether a caller that
-/// read exactly `cap` bytes ended on a newline (legal) or just happened
-/// to saturate the `take` adapter (illegal).
-///
-/// `buf` is NOT cleared — the caller is responsible for clearing it if
-/// they want a fresh line. This lets the function be used both with a
-/// scratch buffer (clear-before-call, the envelope path) and a
-/// freshly-allocated one (the handshake path).
-async fn read_bounded_line<R>(
-    reader: &mut R,
-    cap: usize,
-    buf: &mut Vec<u8>,
-) -> std::io::Result<BoundedReadOutcome>
-where
-    R: AsyncBufRead + Unpin,
-{
-    let before = buf.len();
-    // `take` returns a new adapter that yields at most `cap + 1` bytes.
-    // We reborrow `reader` so the adapter does not take ownership.
-    let limit = cap as u64 + 1;
-    let mut limited = reader.take(limit);
-    let n = limited.read_until(b'\n', buf).await?;
-    if n == 0 {
-        return Ok(BoundedReadOutcome::Eof);
-    }
-    let ends_with_newline = buf.get(before + n - 1).copied() == Some(b'\n');
-    let payload_len = n; // bytes appended by this call
-
-    // Three terminal conditions after a non-zero read:
-    //   1. line_len <= cap AND newline-terminated → legal Line
-    //   2. line_len  > cap (adapter yielded cap + 1) → Overflow (DoS signal)
-    //   3. line_len <= cap AND no newline → underlying reader hit EOF mid-line. Treat as unexpected
-    //      close; the handle is not reusable either way so we surface it as Eof (poisons the handle
-    //      the same way a clean EOF would) rather than as a misleading "exceeded cap" error.
-    if ends_with_newline && payload_len <= cap {
-        // payload_len >= 1 here (n != 0 and the last byte is `\n`), so
-        // `payload_len - 1` (body without the newline) cannot underflow.
-        return Ok(BoundedReadOutcome::Line {
-            body_len: payload_len - 1,
-        });
-    }
-    if payload_len > cap {
-        // The adapter was allowed cap + 1 bytes and consumed them all —
-        // the underlying line is strictly longer than cap. Report as
-        // overflow; do not attempt to read more.
-        return Ok(BoundedReadOutcome::Overflow {
-            observed: payload_len,
-        });
-    }
-    // payload_len <= cap and no trailing newline → partial read with
-    // EOF. Report as Eof — the caller treats it the same as a clean
-    // close (poison the handle, fail with PluginClosed).
-    Ok(BoundedReadOutcome::Eof)
-}
-
-/// Classification of a [`ProcessSandbox::try_dispatch`] failure.
-///
-/// `dispatch_envelope` uses this to decide whether a failed attempt is
-/// eligible for a single respawn-and-retry. Only failures that demonstrably
-/// occurred **before** any envelope bytes landed on a running plugin
-/// process are retried — any other failure (cancellation, timeout,
-/// mid-round-trip transport error, protocol violation) is terminal because
-/// a retry would risk re-invoking a non-idempotent action on the plugin
-/// side. See `dispatch_envelope` docs and the #257 review for the full
-/// rationale.
-#[derive(Debug)]
-enum TryDispatchError {
-    /// The first attempt observed a stale handle (plugin crashed or
-    /// exited between calls). No envelope bytes reached a running plugin
-    /// process, so the outer [`ProcessSandbox::dispatch_envelope`] is
-    /// safe to respawn and retry exactly once.
-    Respawnable(ActionError),
-    /// Terminal for this dispatch — either cancellation, timeout, a
-    /// mid-round-trip transport error, a protocol violation, or a spawn
-    /// failure. Must NOT be retried silently; the engine's higher-level
-    /// retry policy (if any) remains free to retry externally, but the
-    /// sandbox itself has to surface the error as-is so cancellation and
-    /// fatal classifications round-trip correctly.
-    Terminal(ActionError),
-}
-
-impl TryDispatchError {
-    /// Classify a transport-layer [`SandboxError`], given whether the
-    /// outbound envelope was already written to the plugin (`sent`).
-    ///
-    /// [`SandboxError::PluginClosed`] is respawn-eligible **only** when
-    /// `sent == false` — i.e. no envelope bytes reached a running plugin
-    /// for this attempt. Once `send_envelope` has succeeded the plugin may
-    /// have received and begun executing a non-idempotent action before
-    /// dying; resending it on a fresh process would double-execute. The
-    /// original #257 logic treated every `PluginClosed` as safe to
-    /// respawn, conflating "stale handle on entry" with "EOF observed by
-    /// `recv` after a successful send"; the `sent` bit separates them.
-    /// Every other variant is terminal regardless.
-    fn from_sandbox_error_after_send(err: SandboxError, sent: bool) -> Self {
-        let respawnable = !sent && matches!(err, SandboxError::PluginClosed);
-        let action_err = sandbox_error_to_action_error(err);
-        if respawnable {
-            Self::Respawnable(action_err)
-        } else {
-            Self::Terminal(action_err)
-        }
-    }
-
-    /// `true` if [`ProcessSandbox::dispatch_envelope`] is allowed to
-    /// respawn the plugin and retry this envelope once.
-    #[cfg(test)]
-    fn is_respawnable(&self) -> bool {
-        matches!(self, Self::Respawnable(_))
-    }
-
-    /// Unwrap the carried [`ActionError`] once the dispatch-level retry
-    /// decision has been made.
-    fn into_action_error(self) -> ActionError {
-        match self {
-            Self::Respawnable(err) | Self::Terminal(err) => err,
-        }
-    }
-}
-
-/// Outcome of a [`race_cancel_timeout`] call.
-#[derive(Debug, PartialEq, Eq)]
-enum RaceOutcome<T> {
-    /// The inner future produced a value within the deadline and before
-    /// the cancellation token fired.
-    Ready(T),
-    /// The wall-clock deadline elapsed first. The inner future was
-    /// dropped mid-flight; callers must assume its side effects are
-    /// partially applied (writes may have reached the peer).
-    Timeout,
-    /// The optional cancellation token fired first. Same partial-side-effect
-    /// caveat as `Timeout` applies — the race only wins a snapshot of
-    /// progress, not a clean rollback.
-    Cancelled,
-}
-
-/// Race `fut` against a wall-clock deadline and (optionally) a
-/// [`CancellationToken`]. Used by [`ProcessSandbox::try_dispatch`] to
-/// honour both the per-call plugin timeout and the engine-wide
-/// cancellation contract — see #257.
-///
-/// When `cancel` is `Some`, the select is `biased` so a token that is
-/// already cancelled on entry is observed on the very first poll, before
-/// the roundtrip future is driven. When `cancel` is `None` (public helpers
-/// like `invoke` / `get_metadata` that run outside an execution context)
-/// the helper degrades to a plain `tokio::time::timeout`.
-async fn race_cancel_timeout<F, T>(
-    fut: F,
-    timeout: Duration,
-    cancel: Option<&CancellationToken>,
-) -> RaceOutcome<T>
-where
-    F: Future<Output = T>,
-{
-    let timed = tokio::time::timeout(timeout, fut);
-    match cancel {
-        Some(token) => {
-            tokio::select! {
-                biased;
-                () = token.cancelled() => RaceOutcome::Cancelled,
-                r = timed => match r {
-                    Ok(v) => RaceOutcome::Ready(v),
-                    Err(_) => RaceOutcome::Timeout,
-                },
-            }
-        },
-        None => match timed.await {
-            Ok(v) => RaceOutcome::Ready(v),
-            Err(_) => RaceOutcome::Timeout,
-        },
-    }
-}
-
-impl ProcessSandbox {
-    /// Create a new process sandbox for a plugin binary.
-    #[must_use]
-    pub fn new(binary: PathBuf, timeout: Duration) -> Self {
-        Self {
-            binary,
-            timeout,
-            linux_rlimits: LinuxRlimits::default(),
-            handle: Mutex::new(None),
-            next_id: AtomicU64::new(1),
-        }
-    }
-
-    /// Reserve the next monotonic correlation id for an outbound
-    /// envelope (#285). Uses `Relaxed` ordering — id allocation has no
-    /// happens-before requirement against any other memory op; we only
-    /// need uniqueness, which `fetch_add` guarantees regardless of
-    /// ordering.
-    fn next_envelope_id(&self) -> u64 {
-        self.next_id.fetch_add(1, Ordering::Relaxed)
-    }
-
-    /// Override Linux child-process rlimits for this sandbox instance.
-    ///
-    /// On non-Linux platforms the provided limits are ignored.
-    #[must_use]
-    pub fn with_linux_rlimits(mut self, linux_rlimits: LinuxRlimits) -> Self {
-        self.linux_rlimits = linux_rlimits;
-        self
-    }
-
-    /// Invoke an action and return the plugin's response envelope.
-    pub(crate) async fn call_action(
-        &self,
-        action_key: &str,
-        input: serde_json::Value,
-        cancel: Option<&CancellationToken>,
-    ) -> Result<PluginToHost, ActionError> {
-        let request = HostToPlugin::ActionInvoke {
-            id: self.next_envelope_id(),
-            action_key: action_key.to_owned(),
-            input,
-        };
-        self.dispatch_envelope(request, cancel).await
-    }
-
-    /// Query plugin metadata via a `MetadataRequest` envelope.
-    pub async fn get_metadata(&self) -> Result<PluginToHost, ActionError> {
-        let request = HostToPlugin::MetadataRequest {
-            id: self.next_envelope_id(),
-        };
-        self.dispatch_envelope(request, None).await
-    }
-
-    /// Query plugin metadata and return the response as raw JSON bytes,
-    /// **before** attempting the strongly-typed `PluginToHost` deserialize.
-    ///
-    /// Purpose: let the caller check `protocol_version` before the typed
-    /// parse runs. If a v2 plugin sends its old envelope shape (flat
-    /// `plugin_key` / `plugin_version` fields, no `manifest`) against a
-    /// v3 host, this method returns the raw bytes; the caller parses to
-    /// `serde_json::Value` for the version check, and then — if the
-    /// version matches — re-parses the same bytes into `PluginToHost`.
-    ///
-    /// Returning bytes rather than `serde_json::Value` preserves the
-    /// zero-copy / borrowed-`&str` deserialize path that some `Deserialize`
-    /// impls rely on (notably `domain_key::Key<T>` and therefore
-    /// `PluginKey`), which would otherwise fail with
-    /// `"expected a borrowed string"` when going through
-    /// `serde_json::from_value`.
-    ///
-    /// Used by `discover_directory` in the `discovery` module (private path).
-    pub async fn get_metadata_raw(&self) -> Result<Vec<u8>, ActionError> {
-        let request = HostToPlugin::MetadataRequest {
-            id: self.next_envelope_id(),
-        };
-        self.dispatch_envelope_bytes(request, None).await
-    }
-
-    /// High-level action invocation for host code outside the engine flow
-    /// (diagnostics, examples, integration tests, ad-hoc CLI invocations).
-    ///
-    /// Sends an `ActionInvoke` envelope to the (possibly already-spawned)
-    /// plugin process, awaits the matching `ActionResult*` envelope, and
-    /// returns the unwrapped output value.
-    ///
-    /// Production action execution in the engine still goes through the
-    /// `SandboxRunner::execute` trait method, which wraps cancellation,
-    /// metadata plumbing, and integration with `ActionRuntime`.
-    pub async fn invoke(
-        &self,
-        action_key: &str,
-        input: serde_json::Value,
-    ) -> Result<serde_json::Value, ActionError> {
-        let envelope = self.call_action(action_key, input, None).await?;
-        match envelope {
-            PluginToHost::ActionResultOk { output, .. } => Ok(output),
-            PluginToHost::ActionResultError {
-                code,
-                message,
-                retryable,
-                ..
-            } => {
-                let msg = sanitize_plugin_string(&format!("{code}: {message}"));
-                if retryable {
-                    Err(ActionError::retryable(msg))
-                } else {
-                    Err(ActionError::fatal(msg))
-                }
-            },
-            other => Err(ActionError::fatal(format!(
-                "plugin returned unexpected envelope (expected ActionResult*, got {})",
-                envelope_kind(&other)
-            ))),
-        }
-    }
-
-    /// Core long-lived dispatch. Reuses the cached [`PluginHandle`] if any,
-    /// spawns fresh otherwise.
-    ///
-    /// Retry policy (review feedback on #257): we retry **only** when the
-    /// first attempt failed with [`TryDispatchError::Respawnable`] —
-    /// concretely, [`SandboxError::PluginClosed`] (stale handle, plugin
-    /// crashed between calls). Every other failure is terminal on the
-    /// first attempt:
-    ///
-    /// - `Cancelled` is returned as-is to honour the engine's cancellation contract (re-sending
-    ///   could duplicate a non-idempotent action after the caller gave up — see #257 review).
-    /// - Timeout, protocol violations (`ResponseIdMismatch`, `PluginLineTooLarge`,
-    ///   `HandshakeLineTooLarge`, `TransportPoisoned`, `MalformedEnvelope`), and I/O failures
-    ///   mid-round-trip are treated as terminal because the stream position is undefined after a
-    ///   partial write; a blind retry would risk duplicate side-effects on the plugin process.
-    async fn dispatch_envelope(
-        &self,
-        envelope: HostToPlugin,
-        cancel: Option<&CancellationToken>,
-    ) -> Result<PluginToHost, ActionError> {
-        match self.try_dispatch(envelope.clone(), cancel).await {
-            Ok(response) => Ok(response),
-            Err(TryDispatchError::Respawnable(_)) => {
-                // Stale handle — the plugin crashed or exited before we
-                // sent this envelope. Respawning and resending is safe
-                // because no bytes reached a running plugin process.
-                *self.handle.lock().await = None;
-                match self.try_dispatch(envelope, cancel).await {
-                    Ok(response) => Ok(response),
-                    Err(err) => Err(err.into_action_error()),
-                }
-            },
-            Err(err) => Err(err.into_action_error()),
-        }
-    }
-
-    /// Variant of [`dispatch_envelope`] that returns the response as raw
-    /// JSON bytes (no strongly-typed `PluginToHost` parse).
-    ///
-    /// Used by [`get_metadata_raw`](Self::get_metadata_raw): lets the caller
-    /// inspect `protocol_version` before committing to the typed parse,
-    /// which otherwise fails with a confusing "missing field" message on
-    /// a version-mismatched envelope.
-    async fn dispatch_envelope_bytes(
-        &self,
-        envelope: HostToPlugin,
-        cancel: Option<&CancellationToken>,
-    ) -> Result<Vec<u8>, ActionError> {
-        match self.try_dispatch_bytes(envelope.clone(), cancel).await {
-            Ok(response) => Ok(response),
-            Err(TryDispatchError::Respawnable(_)) => {
-                *self.handle.lock().await = None;
-                match self.try_dispatch_bytes(envelope, cancel).await {
-                    Ok(response) => Ok(response),
-                    Err(err) => Err(err.into_action_error()),
-                }
-            },
-            Err(err) => Err(err.into_action_error()),
-        }
-    }
-
-    async fn try_dispatch(
-        &self,
-        envelope: HostToPlugin,
-        cancel: Option<&CancellationToken>,
-    ) -> Result<PluginToHost, TryDispatchError> {
-        let mut guard = self.handle.lock().await;
-        if guard.is_none() {
-            // Spawn failure before any bytes are on the wire — no
-            // side-effect risk, but respawn-retrying a binary that just
-            // failed to start almost never helps. Classify as terminal.
-            let handle = self
-                .spawn_and_dial()
-                .await
-                .map_err(TryDispatchError::Terminal)?;
-            *guard = Some(handle);
-        }
-        let Some(handle) = guard.as_mut() else {
-            // Unreachable in practice: we just set `*guard = Some(..)`
-            // above. Prefer a typed fatal error over `expect(..)` so a
-            // hypothetical logic bug surfaces through the engine's
-            // standard error path instead of panicking inside the
-            // sandbox lock.
-            return Err(TryDispatchError::Terminal(ActionError::fatal(
-                "process sandbox handle missing after successful spawn",
-            )));
-        };
-
-        // Round-trip the envelope with a per-call timeout AND a race
-        // against the engine's cancellation token. Without the cancel
-        // arm, a cancelled workflow would have to wait out `self.timeout`
-        // on a hung or slow plugin before the engine could reclaim the
-        // slot — see #257.
-        let envelope_tag = match &envelope {
-            HostToPlugin::ActionInvoke { .. } => "action_invoke",
-            HostToPlugin::MetadataRequest { .. } => "metadata_request",
-            _ => "other",
-        };
-        // Remember the outbound correlation id so we can validate the
-        // response echoes it back (#285).
-        let expected_id = request_id(&envelope);
-
-        // The `bool` is `sent`: `false` if the failure happened at/before
-        // the write (no bytes reached the plugin), `true` once the write
-        // succeeded (the action may have run — no blind respawn-retry).
-        let roundtrip = async {
-            handle
-                .send_envelope(&envelope)
-                .await
-                .map_err(|e| (false, e))?;
-            handle.recv_envelope().await.map_err(|e| (true, e))
-        };
-
-        let outcome = race_cancel_timeout(roundtrip, self.timeout, cancel).await;
-
-        match outcome {
-            RaceOutcome::Ready(Ok(response)) => {
-                if let (Some(expected), Some(got)) = (expected_id, response_id(&response))
-                    && expected != got
-                {
-                    tracing::warn!(
-                        plugin = %self.binary.display(),
-                        envelope = %envelope_tag,
-                        expected,
-                        got,
-                        "plugin response id mismatch — poisoning handle",
-                    );
-                    *guard = None;
-                    // Protocol violation — must not retry. A stale
-                    // response on a fresh connection is indistinguishable
-                    // from an attacker replaying a prior reply, and a
-                    // retry on a fresh handle would still send a fresh
-                    // request the plugin may already have processed.
-                    return Err(TryDispatchError::Terminal(sandbox_error_to_action_error(
-                        SandboxError::ResponseIdMismatch { expected, got },
-                    )));
-                }
-                Ok(response)
-            },
-            RaceOutcome::Ready(Err((sent, sandbox_err))) => {
-                // Transport/protocol error — invalidate the handle so the
-                // next call respawns. Log PluginLineTooLarge at warn so it
-                // shows up in security dashboards.
-                if matches!(
-                    sandbox_err,
-                    SandboxError::PluginLineTooLarge { .. }
-                        | SandboxError::HandshakeLineTooLarge { .. }
-                        | SandboxError::TransportPoisoned
-                ) {
-                    tracing::warn!(
-                        plugin = %self.binary.display(),
-                        envelope = %envelope_tag,
-                        error = %sandbox_err,
-                        "plugin transport poisoned — clearing handle and forcing respawn",
-                    );
-                }
-                *guard = None;
-                Err(TryDispatchError::from_sandbox_error_after_send(
-                    sandbox_err,
-                    sent,
-                ))
-            },
-            RaceOutcome::Timeout => {
-                // Timeout — also invalidate; we don't know if the plugin is
-                // still processing and we can't safely reuse the connection.
-                // Classified as terminal: silently retrying after the
-                // engine already gave up on this call would risk the
-                // plugin running the action twice (#257 review).
-                *guard = None;
-                Err(TryDispatchError::Terminal(ActionError::retryable(format!(
-                    "plugin {} timed out on {envelope_tag} after {:?}",
-                    self.binary.display(),
-                    self.timeout
-                ))))
-            },
-            RaceOutcome::Cancelled => {
-                // Cancellation observed mid-round-trip. We may have
-                // written part of an envelope to the plugin; the stream
-                // position is undefined, so drop the handle and force a
-                // respawn on the next call. Surface as `ActionError::Cancelled`
-                // so the engine honours the standard cancellation path —
-                // and crucially do NOT retry (would duplicate work the
-                // engine already asked us to abort; see #257 review).
-                *guard = None;
-                tracing::debug!(
-                    plugin = %self.binary.display(),
-                    envelope = %envelope_tag,
-                    "plugin dispatch cancelled via CancellationToken; clearing handle",
-                );
-                Err(TryDispatchError::Terminal(ActionError::Cancelled))
-            },
-        }
-    }
-
-    /// Variant of [`try_dispatch`](Self::try_dispatch) that reads the
-    /// plugin's response as raw JSON bytes instead of parsing to
-    /// `PluginToHost`.
-    ///
-    /// Identical transport / cancel / timeout semantics; only the inbound
-    /// parse step differs. Correlation-id matching happens by parsing the
-    /// bytes once into `serde_json::Value` solely to pull `.id`; the
-    /// caller re-parses the bytes into the target type (see
-    /// `dispatch_envelope_bytes` callers).
-    async fn try_dispatch_bytes(
-        &self,
-        envelope: HostToPlugin,
-        cancel: Option<&CancellationToken>,
-    ) -> Result<Vec<u8>, TryDispatchError> {
-        let mut guard = self.handle.lock().await;
-        if guard.is_none() {
-            let handle = self
-                .spawn_and_dial()
-                .await
-                .map_err(TryDispatchError::Terminal)?;
-            *guard = Some(handle);
-        }
-        let Some(handle) = guard.as_mut() else {
-            return Err(TryDispatchError::Terminal(ActionError::fatal(
-                "process sandbox handle missing after successful spawn",
-            )));
-        };
-
-        let envelope_tag = match &envelope {
-            HostToPlugin::ActionInvoke { .. } => "action_invoke",
-            HostToPlugin::MetadataRequest { .. } => "metadata_request",
-            _ => "other",
-        };
-        let expected_id = request_id(&envelope);
-
-        // `bool` is `sent` (see `try_dispatch`): false at/before write,
-        // true once the write succeeded.
-        let roundtrip = async {
-            handle
-                .send_envelope(&envelope)
-                .await
-                .map_err(|e| (false, e))?;
-            handle.recv_envelope_bytes().await.map_err(|e| (true, e))
-        };
-
-        let outcome = race_cancel_timeout(roundtrip, self.timeout, cancel).await;
-
-        match outcome {
-            RaceOutcome::Ready(Ok(response_bytes)) => {
-                // Cheap reparse to `Value` just for the id check; the
-                // caller will parse the bytes again into the strongly-
-                // typed target. Parsing twice is acceptable because the
-                // metadata-probe path runs once per plugin lifetime.
-                if let Some(expected) = expected_id {
-                    // Use a match rather than `map_err(...)?` so we can
-                    // clear `*guard` on parse failure before returning.
-                    // A `?` return would skip the guard-clear and leave a
-                    // cached-but-poisoned handle for the next call.
-                    let value: serde_json::Value = match serde_json::from_slice(&response_bytes) {
-                        Ok(v) => v,
-                        Err(e) => {
-                            *guard = None;
-                            // Response bytes were received, so the action
-                            // ran — terminal, never a silent resend.
-                            return Err(TryDispatchError::from_sandbox_error_after_send(
-                                SandboxError::MalformedEnvelope(e),
-                                true,
-                            ));
-                        },
-                    };
-                    if let Some(got) = response_id_from_value(&value)
-                        && expected != got
-                    {
-                        tracing::warn!(
-                            plugin = %self.binary.display(),
-                            envelope = %envelope_tag,
-                            expected,
-                            got,
-                            "plugin response id mismatch — poisoning handle",
-                        );
-                        *guard = None;
-                        return Err(TryDispatchError::Terminal(sandbox_error_to_action_error(
-                            SandboxError::ResponseIdMismatch { expected, got },
-                        )));
-                    }
-                }
-                Ok(response_bytes)
-            },
-            RaceOutcome::Ready(Err((sent, sandbox_err))) => {
-                if matches!(
-                    sandbox_err,
-                    SandboxError::PluginLineTooLarge { .. }
-                        | SandboxError::HandshakeLineTooLarge { .. }
-                        | SandboxError::TransportPoisoned
-                ) {
-                    tracing::warn!(
-                        plugin = %self.binary.display(),
-                        envelope = %envelope_tag,
-                        error = %sandbox_err,
-                        "plugin transport poisoned — clearing handle and forcing respawn",
-                    );
-                }
-                *guard = None;
-                Err(TryDispatchError::from_sandbox_error_after_send(
-                    sandbox_err,
-                    sent,
-                ))
-            },
-            RaceOutcome::Timeout => {
-                *guard = None;
-                Err(TryDispatchError::Terminal(ActionError::retryable(format!(
-                    "plugin {} timed out on {envelope_tag} after {:?}",
-                    self.binary.display(),
-                    self.timeout
-                ))))
-            },
-            RaceOutcome::Cancelled => {
-                *guard = None;
-                tracing::debug!(
-                    plugin = %self.binary.display(),
-                    envelope = %envelope_tag,
-                    "plugin dispatch cancelled via CancellationToken; clearing handle",
-                );
-                Err(TryDispatchError::Terminal(ActionError::Cancelled))
-            },
-        }
-    }
-
-    /// Spawn the plugin binary, read and parse its handshake line, dial the
-    /// announced transport, and return a fresh [`PluginHandle`].
-    async fn spawn_and_dial(&self) -> Result<PluginHandle, ActionError> {
-        // #260: host allocates the plugin's socket address up-front and
-        // passes it via env so the child cannot forge a handshake that
-        // redirects the host at a sibling plugin's socket. `socket_dir`
-        // is `Some` on Unix (TempDir that owns the 0700 parent) and
-        // `None` on Windows (named pipes aren't in a filesystem dir).
-        let (expected_addr, kind, socket_dir) = allocate_host_socket_addr()?;
-
-        // `env_clear()` then only the host-controlled transport env: a
-        // Phase-era plugin inherits **no** host environment. A
-        // host-authored, scope-keyed env allowlist returns only with the
-        // broker (ADR-0025 §6), never as a plugin-declared capability.
-        let mut cmd = Command::new(&self.binary);
-        cmd.stdin(std::process::Stdio::null())
-            .stdout(std::process::Stdio::piped())
-            .stderr(std::process::Stdio::piped())
-            .kill_on_drop(true)
-            .env_clear()
-            .env(ENV_SOCKET_ADDR, &expected_addr)
-            .env(ENV_SOCKET_KIND, kind);
-
-        // Apply OS-level hardening in the child between fork() and exec()
-        // (Linux only). The entire Landlock ruleset + rlimit snapshot is
-        // built HERE, pre-fork, on the host thread — every allocation
-        // happens before fork.
-        #[cfg(target_os = "linux")]
-        {
-            let mut prepared = crate::os_sandbox::PreparedSandbox::prepare(self.linux_rlimits)
-                .map_err(|e| ActionError::fatal(format!("sandbox prepare failed: {e}")))?;
-
-            // SAFETY: the closure runs between fork() and exec(). It calls
-            // only setrlimit(2) and landlock_restrict_self(2) on data
-            // allocated above (pre-fork) — no allocation, no serde, no
-            // PathFd::new, no tracing on the success path. This is the
-            // structural fix for the post-fork allocator-lock deadlock
-            // class: a multi-threaded parent may hold the allocator lock
-            // at fork, so the child must not re-enter the allocator.
-            #[expect(
-                unsafe_code,
-                reason = "pre_exec: only setrlimit + landlock_restrict_self on pre-fork-allocated data; no allocation in child"
-            )]
-            unsafe {
-                cmd.pre_exec(move || prepared.apply_in_child());
-            }
-        }
-
-        let mut child = cmd.spawn().map_err(|e| {
-            ActionError::fatal(format!(
-                "failed to spawn plugin {}: {e}",
-                self.binary.display()
-            ))
-        })?;
-
-        // Spawn a background task that drains the plugin's stderr and logs
-        // each line via `tracing`. We do this BEFORE reading the handshake
-        // so that any crash diagnostics the plugin writes during startup
-        // are captured.
-        //
-        // Fire-and-forget: the returned JoinHandle is intentionally
-        // dropped (#270). The task's lifecycle is bounded implicitly via
-        // the chain:
-        //
-        //   cmd.kill_on_drop(true)             (above, ~"kill_on_drop")
-        //     → dropping PluginHandle drops Child
-        //     → Child's Drop terminates the plugin process
-        //       (SIGKILL on Unix, TerminateProcess on Windows — both
-        //       close the child's stderr pipe handle as a side effect)
-        //     → plugin's stderr pipe closes
-        //     → drain_plugin_stderr observes EOF and returns
-        //     → detached task completes and is reaped
-        //
-        // If a future refactor removes `kill_on_drop(true)`, make sure
-        // the replacement shutdown path still reliably terminates the
-        // child so stderr closes and the drainer reaches EOF, and/or
-        // keep an owned JoinHandle that can be aborted explicitly on
-        // ProcessSandbox::drop. The leak risk is not "explicit
-        // shutdown" by itself; it is any shutdown path that leaves the
-        // stderr drainer with no guaranteed completion signal — at
-        // that point tokio's task slab grows unbounded per plugin
-        // spawn.
-        if let Some(stderr) = child.stderr.take() {
-            let plugin_name = self.binary.display().to_string();
-            tokio::spawn(drain_plugin_stderr(stderr, plugin_name));
-        }
-
-        // Read the handshake line from child stdout with a hard timeout
-        // AND a hard length cap — a malicious plugin must not be able to
-        // hold this task blocking on memory growth.
-        let stdout = child
-            .stdout
-            .take()
-            .ok_or_else(|| ActionError::fatal("failed to open plugin stdout"))?;
-        let mut stdout_reader = BufReader::new(stdout);
-        let mut handshake_buf: Vec<u8> = Vec::with_capacity(256);
-
-        let read_result = tokio::time::timeout(HANDSHAKE_TIMEOUT, async {
-            read_bounded_line(&mut stdout_reader, HANDSHAKE_LINE_CAP, &mut handshake_buf).await
-        })
-        .await;
-
-        let outcome = read_result.map_err(|_| {
-            ActionError::fatal(format!(
-                "plugin {} handshake timeout after {HANDSHAKE_TIMEOUT:?}",
-                self.binary.display()
-            ))
-        })?;
-
-        let body_len = match outcome {
-            Ok(BoundedReadOutcome::Line { body_len }) => body_len,
-            Ok(BoundedReadOutcome::Eof) => {
-                return Err(ActionError::fatal(format!(
-                    "plugin {} exited before printing handshake line",
-                    self.binary.display()
-                )));
-            },
-            Ok(BoundedReadOutcome::Overflow { observed }) => {
-                tracing::warn!(
-                    plugin = %self.binary.display(),
-                    limit = HANDSHAKE_LINE_CAP,
-                    observed,
-                    "plugin handshake exceeded cap — refusing to dial",
-                );
-                return Err(sandbox_error_to_action_error(
-                    SandboxError::HandshakeLineTooLarge {
-                        limit: HANDSHAKE_LINE_CAP,
-                        observed,
-                    },
-                ));
-            },
-            Err(e) => {
-                return Err(ActionError::fatal(format!(
-                    "plugin {} handshake read error: {e}",
-                    self.binary.display()
-                )));
-            },
-        };
-
-        // Strip the trailing newline and decode as UTF-8 for the dial
-        // address. We do this AFTER the cap check so we never run UTF-8
-        // validation on an unbounded buffer.
-        let handshake_bytes = &handshake_buf[..body_len];
-        let handshake_line = std::str::from_utf8(handshake_bytes).map_err(|e| {
-            ActionError::fatal(format!(
-                "plugin {} handshake line is not valid UTF-8: {e}",
-                self.binary.display()
-            ))
-        })?;
-
-        let sanitized_handshake = sanitize_plugin_string(handshake_line.trim());
-        tracing::debug!(
-            plugin = %self.binary.display(),
-            handshake = %sanitized_handshake,
-            "plugin handshake received"
-        );
-
-        // #260: validate the announced address against what we pre-allocated
-        // before dialling. A compromised plugin that prints a sibling
-        // plugin's socket path here is rejected, not connected to.
-        if let Err(err) = validate_handshake_addr(handshake_line, &expected_addr, kind) {
-            tracing::warn!(
-                plugin = %self.binary.display(),
-                expected = %expected_addr,
-                error = %err,
-                "plugin handshake address mismatch — refusing to dial",
-            );
-            return Err(sandbox_error_to_action_error(err));
-        }
-
-        // Dial the announced transport.
-        let stream = transport::dial(handshake_line)
-            .await
-            .map_err(|e| ActionError::fatal(format!("plugin transport dial failed: {e}")))?;
-
-        Ok(PluginHandle::new(child, stream, socket_dir))
-    }
-}
-
-/// Allocate a host-controlled socket address for the plugin to bind (#260).
-///
-/// On Unix: creates a tempdir with `0700` permissions (via `tempfile`) and
-/// returns a socket path inside it. Keeping the `TempDir` alive is the
-/// caller's responsibility — it must be stored on the resulting
-/// [`PluginHandle`] so the directory persists for as long as the socket
-/// is in use and is cleaned up on drop.
-///
-/// Prefers `/tmp` when available over the platform temp dir (macOS
-/// defaults to `/var/folders/...` paths that can exceed
-/// [`transport::MAX_UNIX_SOCKET_PATH_BYTES`]). Falls back to the platform
-/// temp dir, and skips any candidate whose resulting socket path would
-/// overflow `sun_path`. This mirrors the plugin-side self-allocation
-/// policy so the host and plugin cannot disagree on what's bindable.
-///
-/// On Windows: generates an unpredictable named-pipe path under
-/// `\\.\pipe\LOCAL\` (session-scoped namespace). The pipe is bound by
-/// the plugin and released when the plugin process exits; no
-/// directory-level cleanup is needed.
-fn allocate_host_socket_addr()
--> Result<(String, &'static str, Option<tempfile::TempDir>), ActionError> {
-    #[cfg(unix)]
-    {
-        use std::{os::unix::ffi::OsStrExt, path::PathBuf};
-
-        // Candidate roots in preference order: short `/tmp` first (macOS
-        // `/var/folders/…` often exceeds the `sun_path` cap), then the
-        // platform temp dir. Deduplicate so `/tmp`-is-temp_dir callers
-        // don't try the same root twice.
-        let mut candidate_roots: Vec<PathBuf> = Vec::new();
-        let short_tmp = PathBuf::from("/tmp");
-        if short_tmp.is_dir() {
-            candidate_roots.push(short_tmp);
-        }
-        let platform_tmp = std::env::temp_dir();
-        if !candidate_roots.iter().any(|root| root == &platform_tmp) {
-            candidate_roots.push(platform_tmp);
-        }
-
-        let mut last_alloc_error: Option<(PathBuf, std::io::Error)> = None;
-        for root in candidate_roots {
-            let dir = match tempfile::Builder::new()
-                .prefix("nebula-plugin-host-")
-                .tempdir_in(&root)
-            {
-                Ok(dir) => dir,
-                Err(e) => {
-                    last_alloc_error = Some((root, e));
-                    continue;
-                },
-            };
-
-            let socket_path = dir.path().join("sock");
-            let socket_path_len = socket_path.as_os_str().as_bytes().len();
-            if socket_path_len > transport::MAX_UNIX_SOCKET_PATH_BYTES {
-                // Drop the oversized tempdir (Drop cleans up) and try
-                // the next candidate root.
-                continue;
-            }
-
-            let addr = socket_path
-                .to_str()
-                .ok_or_else(|| ActionError::fatal("plugin socket tempdir path is not valid UTF-8"))?
-                .to_owned();
-            return Ok((addr, "unix", Some(dir)));
-        }
-
-        if let Some((root, e)) = last_alloc_error {
-            return Err(ActionError::fatal(format!(
-                "failed to allocate plugin socket tempdir in {}: {e}",
-                root.display()
-            )));
-        }
-        Err(ActionError::fatal(format!(
-            "failed to allocate a Unix socket path within {} bytes",
-            transport::MAX_UNIX_SOCKET_PATH_BYTES
-        )))
-    }
-    #[cfg(windows)]
-    {
-        let nonce = uuid::Uuid::new_v4().simple().to_string();
-        let pipe = format!(r"\\.\pipe\LOCAL\nebula-plugin-host-{nonce}");
-        Ok((pipe, "pipe", None))
-    }
-}
-
-/// Verify that the handshake line's `kind|addr` pair matches the
-/// host-allocated values from [`allocate_host_socket_addr`]. Returns
-/// [`SandboxError::HandshakeAddrMismatch`] on any deviation.
-///
-/// Exact string comparison is used — the plugin is expected to echo the
-/// address we passed via `NEBULA_PLUGIN_SOCKET_ADDR` verbatim. An
-/// attacker-controlled plugin that prints a different address (to
-/// redirect the host at a sibling plugin's UDS or pipe) fails this check
-/// and never reaches `transport::dial`.
-fn validate_handshake_addr(
-    handshake_line: &str,
-    expected_addr: &str,
-    expected_kind: &'static str,
-) -> Result<(), SandboxError> {
-    let line = handshake_line.trim();
-    let mut parts = line.splitn(3, '|');
-    let _version = parts.next(); // already checked by the caller's length cap and UTF-8 decode
-    let announced_kind = parts.next().unwrap_or("");
-    let announced_addr = parts.next().unwrap_or("");
-    if announced_kind != expected_kind || announced_addr != expected_addr {
-        let got = if announced_kind.is_empty() && announced_addr.is_empty() {
-            String::from("<malformed handshake>")
-        } else {
-            format!("{announced_kind}|{announced_addr}")
-        };
-        return Err(SandboxError::HandshakeAddrMismatch {
-            expected: format!("{expected_kind}|{expected_addr}"),
-            got,
-        });
-    }
-    Ok(())
-}
-
-/// Convert an internal [`SandboxError`] into the public `ActionError` the
-/// sandbox runner trait returns. Transport-level issues are fatal
-/// (non-retryable) by design: once the plugin has misbehaved on the wire,
-/// the next caller gets a fresh process, not a blind retry on the same
-/// poisoned channel.
-fn sandbox_error_to_action_error(err: SandboxError) -> ActionError {
-    match err {
-        // Retryable: plugin crashed / exited, respawn path is safe.
-        SandboxError::PluginClosed => ActionError::retryable_from(err),
-        // Fatal: DoS / protocol-abuse signals. Do not paper over with retry.
-        SandboxError::PluginLineTooLarge { .. }
-        | SandboxError::HandshakeLineTooLarge { .. }
-        | SandboxError::HandshakeAddrMismatch { .. }
-        | SandboxError::ResponseIdMismatch { .. }
-        | SandboxError::TransportPoisoned
-        | SandboxError::Transport(_)
-        | SandboxError::MalformedEnvelope(_)
-        | SandboxError::HostMalformedEnvelope(_)
-        // Pre-`fork` spawn-time hardening failures. They reach the public
-        // boundary as fatal; a retry on the same misconfigured host would
-        // fail identically.
-        | SandboxError::Landlock(_)
-        | SandboxError::Rlimit(_) => ActionError::fatal_from(err),
-    }
-}
-
-/// Correlation id carried by an outbound host→plugin envelope, if any.
-/// `Shutdown` has no id (it's one-way, no response expected); every
-/// other variant carries a `u64`.
-fn request_id(env: &HostToPlugin) -> Option<u64> {
-    match env {
-        HostToPlugin::ActionInvoke { id, .. }
-        | HostToPlugin::MetadataRequest { id }
-        | HostToPlugin::Cancel { id }
-        | HostToPlugin::RpcResponseOk { id, .. }
-        | HostToPlugin::RpcResponseError { id, .. } => Some(*id),
-        HostToPlugin::Shutdown => None,
-    }
-}
-
-/// Correlation id carried by an inbound plugin→host envelope, if any.
-/// `Log` is one-way and carries no correlation id; every other variant
-/// carries a `u64`.
-fn response_id(env: &PluginToHost) -> Option<u64> {
-    match env {
-        PluginToHost::ActionResultOk { id, .. }
-        | PluginToHost::ActionResultError { id, .. }
-        | PluginToHost::RpcCall { id, .. }
-        | PluginToHost::MetadataResponse { id, .. } => Some(*id),
-        PluginToHost::Log { .. } => None,
-    }
-}
-
-/// Variant of [`response_id`] that works on a raw `serde_json::Value`.
-///
-/// Mirrors the typed path used by `try_dispatch` for the raw-value metadata
-/// probe. Returns `None` when the field is absent (`Log` variant) or when
-/// the value isn't a `u64`.
-fn response_id_from_value(value: &serde_json::Value) -> Option<u64> {
-    value.get("id").and_then(serde_json::Value::as_u64)
-}
-
-#[async_trait]
-impl SandboxRunner for ProcessSandbox {
-    async fn execute(
-        &self,
-        context: SandboxedContext,
-        metadata: &ActionMetadata,
-        input: serde_json::Value,
-    ) -> Result<ActionResult<serde_json::Value>, ActionError> {
-        context.check_cancelled()?;
-
-        let action_key = metadata.base.key.as_str();
-
-        tracing::debug!(
-            action_key = %action_key,
-            plugin = %self.binary.display(),
-            "executing action in process sandbox"
-        );
-
-        let envelope = self
-            .call_action(action_key, input, Some(context.cancellation()))
-            .await?;
-        match envelope {
-            PluginToHost::ActionResultOk { output, .. } => Ok(ActionResult::success(output)),
-            PluginToHost::ActionResultError {
-                code,
-                message,
-                retryable,
-                ..
-            } => {
-                let msg = sanitize_plugin_string(&format!("{code}: {message}"));
-                if retryable {
-                    Err(ActionError::retryable(msg))
-                } else {
-                    Err(ActionError::fatal(msg))
-                }
-            },
-            other => Err(ActionError::fatal(format!(
-                "plugin returned unexpected envelope (expected ActionResult*, got {})",
-                envelope_kind(&other)
-            ))),
-        }
-    }
-}
-
-/// Drop the cached handle on sandbox drop so the child is killed promptly.
-///
-/// `kill_on_drop(true)` on the spawned `Command` handles this at the OS
-/// level — the destructor of `PluginHandle.child` sends SIGKILL. We add no
-/// extra cleanup here; the `Arc<ProcessSandbox>` in the engine's handler
-/// table owns the lifetime.
-impl Drop for ProcessSandbox {
-    fn drop(&mut self) {
-        tracing::debug!(
-            plugin = %self.binary.display(),
-            "ProcessSandbox dropped; plugin child will be killed by kill_on_drop"
-        );
-    }
-}
-
-/// Drain a plugin child's stderr, emitting one `tracing::debug!` event per
-/// line. Returns when the stderr pipe closes (plugin exited) or read errors.
-///
-/// Unlike the envelope path, stderr is diagnostic log output — a
-/// too-long line is truncated silently rather than rejected. What we
-/// MUST still bound is the memory use of the read itself: a plugin that
-/// spams bytes without newlines forever must not grow our buffer forever.
-/// On overflow, we discard the current buffer, drain bytes until the
-/// next newline, and resume. Truncation is logged at debug level so it
-/// stays observable but doesn't pollute warn/error dashboards.
-async fn drain_plugin_stderr(stderr: tokio::process::ChildStderr, plugin_name: String) {
-    let mut reader = BufReader::new(stderr);
-    let mut line_buf: Vec<u8> = Vec::with_capacity(256);
-    loop {
-        line_buf.clear();
-        match read_bounded_line(&mut reader, STDERR_LINE_CAP, &mut line_buf).await {
-            Ok(BoundedReadOutcome::Line { body_len }) => {
-                let body = &line_buf[..body_len];
-                let as_str = String::from_utf8_lossy(body);
-                let sanitized = sanitize_plugin_string(as_str.trim());
-                tracing::debug!(
-                    plugin = %plugin_name,
-                    stderr = %sanitized,
-                    "plugin stderr"
-                );
-            },
-            Ok(BoundedReadOutcome::Eof) => return,
-            Ok(BoundedReadOutcome::Overflow { observed }) => {
-                tracing::debug!(
-                    plugin = %plugin_name,
-                    observed,
-                    limit = STDERR_LINE_CAP,
-                    "plugin stderr line exceeded cap — discarding and resynchronising to next newline",
-                );
-                // Discard rest of the oversized line until next newline or
-                // EOF. Returning `false` from the helper means the stderr
-                // stream has closed — exit the drain loop entirely.
-                if !drop_until_newline(&mut reader).await {
-                    return;
-                }
-            },
-            Err(_) => return,
-        }
-    }
-}
-
-/// Read and discard bytes from `reader` until a newline is encountered
-/// or the stream ends. Returns `true` if a newline was found (drain loop
-/// should continue), `false` if the stream closed / errored (drain loop
-/// should exit).
-///
-/// Bounds memory to a single small scratch buffer regardless of how
-/// many bytes the plugin emits before the next newline. Used to recover
-/// from an oversized stderr line without growing the line buffer past
-/// `STDERR_LINE_CAP`.
-async fn drop_until_newline<R>(reader: &mut R) -> bool
-where
-    R: AsyncBufRead + Unpin,
-{
-    loop {
-        let chunk = match reader.fill_buf().await {
-            Ok(chunk) => chunk,
-            Err(_) => return false,
-        };
-        if chunk.is_empty() {
-            return false;
-        }
-        if let Some(idx) = chunk.iter().position(|&b| b == b'\n') {
-            reader.consume(idx + 1);
-            return true;
-        }
-        let len = chunk.len();
-        reader.consume(len);
-    }
-}
-
-fn envelope_kind(env: &PluginToHost) -> &'static str {
-    match env {
-        PluginToHost::ActionResultOk { .. } => "action_result_ok",
-        PluginToHost::ActionResultError { .. } => "action_result_error",
-        PluginToHost::RpcCall { .. } => "rpc_call",
-        PluginToHost::Log { .. } => "log",
-        PluginToHost::MetadataResponse { .. } => "metadata_response",
-    }
-}
-
-fn sanitize_plugin_string(s: &str) -> String {
-    s.chars()
-        .take(1024)
-        .map(|c| if c.is_control() && c != '\n' { '?' } else { c })
-        .collect()
-}
-
-#[cfg(test)]
-mod tests {
-    //! Unit tests for the bounded plugin transport line reader and the
-    //! poison invariant.
-    //!
-    //! These tests deliberately target the pure helper
-    //! [`read_bounded_line`] rather than spinning up a real plugin
-    //! process. The helper is what actually enforces the cap and drives
-    //! both the handshake and envelope paths in production, so covering
-    //! it exhaustively at this layer gives the whole crate its security
-    //! guarantee.
-    //!
-    //! We also test `recv_envelope` / `send_envelope` poisoning by
-    //! constructing a `PluginHandle`-shaped fixture via
-    //! `tokio::io::duplex` and a test-only constructor. The real
-    //! `PluginHandle` wraps a `PluginStream`, so for fixture purposes we
-    //! introduce a separate `TestHandle` struct that mirrors the same
-    //! fields and methods with the duplex-backed types — any divergence
-    //! from the production `PluginHandle` would be caught by the fact
-    //! that both call into the same `read_bounded_line` primitive.
-
-    use tokio::io::{AsyncWriteExt, BufReader as TokioBufReader, duplex};
-
-    use super::*;
-
-    // ---- read_bounded_line primitive ---------------------------------
-
-    #[tokio::test]
-    async fn read_bounded_line_accepts_short_line() {
-        let data: &[u8] = b"hello\n";
-        let mut reader = TokioBufReader::new(data);
-        let mut buf = Vec::new();
-        let outcome = read_bounded_line(&mut reader, 1024, &mut buf)
-            .await
-            .unwrap();
-        assert_eq!(outcome, BoundedReadOutcome::Line { body_len: 5 });
-        assert_eq!(buf, b"hello\n");
-    }
-
-    #[tokio::test]
-    async fn read_bounded_line_accepts_exactly_cap_bytes() {
-        // cap = 8, line is "aaaaaaa\n" → exactly 8 bytes including newline.
-        // Guards the off-by-one in `take(cap + 1)`.
-        let data: &[u8] = b"aaaaaaa\n";
-        assert_eq!(data.len(), 8);
-        let mut reader = TokioBufReader::new(data);
-        let mut buf = Vec::new();
-        let outcome = read_bounded_line(&mut reader, 8, &mut buf).await.unwrap();
-        assert_eq!(outcome, BoundedReadOutcome::Line { body_len: 7 });
-    }
-
-    #[tokio::test]
-    async fn read_bounded_line_rejects_one_byte_over_cap() {
-        // cap = 4, line is 5 bytes before newline → overflow.
-        let data: &[u8] = b"aaaaa\n";
-        let mut reader = TokioBufReader::new(data);
-        let mut buf = Vec::new();
-        let outcome = read_bounded_line(&mut reader, 4, &mut buf).await.unwrap();
-        match outcome {
-            BoundedReadOutcome::Overflow { observed } => assert_eq!(observed, 5),
-            other => panic!("expected Overflow, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn read_bounded_line_rejects_newline_starved_stream() {
-        // 1 MiB + 1 of 'A', no newline.
-        let mut data = vec![b'A'; 1024 * 1024 + 1];
-        data.push(b'A');
-        let mut reader = TokioBufReader::new(&data[..]);
-        let mut buf = Vec::new();
-        let outcome = read_bounded_line(&mut reader, 1024 * 1024, &mut buf)
-            .await
-            .unwrap();
-        match outcome {
-            BoundedReadOutcome::Overflow { observed } => {
-                // We should have read exactly cap + 1 bytes before giving up.
-                assert_eq!(observed, 1024 * 1024 + 1);
-            },
-            other => panic!("expected Overflow, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn read_bounded_line_reports_eof() {
-        let data: &[u8] = b"";
-        let mut reader = TokioBufReader::new(data);
-        let mut buf = Vec::new();
-        let outcome = read_bounded_line(&mut reader, 1024, &mut buf)
-            .await
-            .unwrap();
-        assert_eq!(outcome, BoundedReadOutcome::Eof);
-    }
-
-    #[tokio::test]
-    async fn read_bounded_line_partial_line_then_eof_is_reported_as_eof() {
-        // Some bytes, no newline, stream ends. Must be reported as Eof,
-        // NOT as Overflow — a short unterminated tail is a clean-close
-        // edge case, not a DoS signal, and mis-classifying it would
-        // flood security dashboards with false positives whenever a
-        // plugin crashes mid-envelope.
-        let data: &[u8] = b"partial";
-        let mut reader = TokioBufReader::new(data);
-        let mut buf = Vec::new();
-        let outcome = read_bounded_line(&mut reader, 1024, &mut buf)
-            .await
-            .unwrap();
-        assert_eq!(outcome, BoundedReadOutcome::Eof);
-    }
-
-    #[tokio::test]
-    async fn read_bounded_line_successive_lines_from_same_reader() {
-        // Two complete lines back-to-back.
-        let data: &[u8] = b"first\nsecond\n";
-        let mut reader = TokioBufReader::new(data);
-        let mut buf = Vec::new();
-
-        let out1 = read_bounded_line(&mut reader, 1024, &mut buf)
-            .await
-            .unwrap();
-        assert_eq!(out1, BoundedReadOutcome::Line { body_len: 5 });
-        assert_eq!(&buf[..6], b"first\n");
-
-        buf.clear();
-        let out2 = read_bounded_line(&mut reader, 1024, &mut buf)
-            .await
-            .unwrap();
-        assert_eq!(out2, BoundedReadOutcome::Line { body_len: 6 });
-        assert_eq!(&buf[..7], b"second\n");
-    }
-
-    // ---- PluginHandle-shaped fixture for poison tests ----------------
-
-    /// Test double for [`PluginHandle`] that uses an in-memory
-    /// [`tokio::io::duplex`] pair instead of a real plugin transport.
-    ///
-    /// This duplicates the recv/send logic from `PluginHandle` faithfully
-    /// enough to exercise the poisoning invariant end-to-end without the
-    /// real `PluginStream` / `Child` types. Any logic drift between this
-    /// and the real handle would show up as the real handle not calling
-    /// `read_bounded_line` — trivially spotted in code review.
-    struct TestHandle {
-        reader: TokioBufReader<tokio::io::DuplexStream>,
-        writer: tokio::io::DuplexStream,
-        line_buf: Vec<u8>,
-        poisoned: bool,
-    }
-
-    impl TestHandle {
-        fn new(reader_side: tokio::io::DuplexStream, writer_side: tokio::io::DuplexStream) -> Self {
-            Self {
-                reader: TokioBufReader::new(reader_side),
-                writer: writer_side,
-                line_buf: Vec::with_capacity(64),
-                poisoned: false,
-            }
-        }
-
-        async fn recv_envelope_capped(&mut self, cap: usize) -> Result<Vec<u8>, SandboxError> {
-            if self.poisoned {
-                return Err(SandboxError::TransportPoisoned);
-            }
-            self.line_buf.clear();
-            match read_bounded_line(&mut self.reader, cap, &mut self.line_buf).await {
-                Ok(BoundedReadOutcome::Line { body_len }) => Ok(self.line_buf[..body_len].to_vec()),
-                Ok(BoundedReadOutcome::Eof) => {
-                    self.poisoned = true;
-                    Err(SandboxError::PluginClosed)
-                },
-                Ok(BoundedReadOutcome::Overflow { observed }) => {
-                    self.poisoned = true;
-                    Err(SandboxError::PluginLineTooLarge {
-                        limit: cap,
-                        observed,
-                    })
-                },
-                Err(e) => {
-                    self.poisoned = true;
-                    Err(SandboxError::Transport(e))
-                },
-            }
-        }
-
-        async fn send_line(&mut self, line: &[u8]) -> Result<(), SandboxError> {
-            if self.poisoned {
-                return Err(SandboxError::TransportPoisoned);
-            }
-            if let Err(e) = self.writer.write_all(line).await {
-                self.poisoned = true;
-                return Err(SandboxError::Transport(e));
-            }
-            if let Err(e) = self.writer.flush().await {
-                self.poisoned = true;
-                return Err(SandboxError::Transport(e));
-            }
-            Ok(())
-        }
-    }
-
-    #[tokio::test]
-    async fn oversized_envelope_poisons_handle_and_blocks_subsequent_reads() {
-        // This test is the named regression for #316: a malicious plugin
-        // emits > ENVELOPE_LINE_CAP bytes on a single line, the read
-        // fails with PluginLineTooLarge, and the handle is marked
-        // unusable so any further recv_envelope call short-circuits
-        // with TransportPoisoned instead of reading more bytes.
-        let cap = 64; // small cap so the test is fast
-        let (mut fake_plugin, host_side) = duplex(cap * 4);
-        let (host_side_a, _host_side_b) = duplex(8); // dummy reverse side
-        drop(_host_side_b);
-
-        // Plugin side emits cap + 10 bytes of 'X', no newline.
-        let mut payload = vec![b'X'; cap + 10];
-        payload.push(b'\n'); // newline exists but too far out
-        fake_plugin.write_all(&payload).await.unwrap();
-        fake_plugin.flush().await.unwrap();
-
-        let mut handle = TestHandle::new(host_side, host_side_a);
-
-        // First recv: the cap is breached, typed error, handle poisoned.
-        let err = handle.recv_envelope_capped(cap).await.unwrap_err();
-        match err {
-            SandboxError::PluginLineTooLarge { limit, observed } => {
-                assert_eq!(limit, cap);
-                assert!(
-                    observed > cap,
-                    "observed={observed} must exceed cap={cap} to count as overflow",
-                );
-            },
-            other => panic!("expected PluginLineTooLarge, got {other:?}"),
-        }
-        assert!(
-            handle.poisoned,
-            "handle must be marked poisoned after overflow"
-        );
-
-        // Second recv: must short-circuit with TransportPoisoned. It must
-        // NOT read more bytes from the plugin side. This is the
-        // "connection invalidation" invariant — verified in code, not
-        // via comments.
-        let err2 = handle.recv_envelope_capped(cap).await.unwrap_err();
-        assert!(
-            matches!(err2, SandboxError::TransportPoisoned),
-            "expected TransportPoisoned on second recv, got {err2:?}",
-        );
-
-        // Third recv: still poisoned.
-        let err3 = handle.recv_envelope_capped(cap).await.unwrap_err();
-        assert!(matches!(err3, SandboxError::TransportPoisoned));
-
-        // Writes must also fail on a poisoned handle.
-        let send_err = handle.send_line(b"hello\n").await.unwrap_err();
-        assert!(matches!(send_err, SandboxError::TransportPoisoned));
-    }
-
-    #[tokio::test]
-    async fn exact_cap_envelope_is_accepted_and_does_not_poison() {
-        // Boundary: a line of exactly `cap` bytes including the newline
-        // must succeed AND leave the handle reusable. This guards the
-        // off-by-one in the `+ 1` inside read_bounded_line.
-        let cap = 16;
-        let (mut fake_plugin, host_side) = duplex(1024);
-        let (host_side_a, _host_side_b) = duplex(8);
-        drop(_host_side_b);
-
-        // 15 bytes of 'Y' + 1 newline = 16 bytes = exactly cap.
-        let mut payload = vec![b'Y'; cap - 1];
-        payload.push(b'\n');
-        assert_eq!(payload.len(), cap);
-        fake_plugin.write_all(&payload).await.unwrap();
-        fake_plugin.flush().await.unwrap();
-
-        let mut handle = TestHandle::new(host_side, host_side_a);
-        let body = handle.recv_envelope_capped(cap).await.unwrap();
-        assert_eq!(body.len(), cap - 1);
-        assert!(body.iter().all(|&b| b == b'Y'));
-        assert!(
-            !handle.poisoned,
-            "exact-cap read must NOT poison the handle"
-        );
-    }
-
-    #[tokio::test]
-    async fn eof_poisons_handle() {
-        // Plugin closed transport without sending anything → PluginClosed
-        // AND handle poisoned (an EOF is terminal by definition).
-        let (fake_plugin, host_side) = duplex(64);
-        drop(fake_plugin); // close the plugin side immediately
-
-        let (host_side_a, _host_side_b) = duplex(8);
-        drop(_host_side_b);
-
-        let mut handle = TestHandle::new(host_side, host_side_a);
-        let err = handle.recv_envelope_capped(64).await.unwrap_err();
-        assert!(matches!(err, SandboxError::PluginClosed));
-        assert!(handle.poisoned);
-
-        let err2 = handle.recv_envelope_capped(64).await.unwrap_err();
-        assert!(matches!(err2, SandboxError::TransportPoisoned));
-    }
-
-    #[tokio::test]
-    async fn drop_until_newline_keeps_bytes_after_newline() {
-        let data: &[u8] = b"oversized-line\nnext-line\n";
-        let mut reader = TokioBufReader::new(data);
-
-        assert!(drop_until_newline(&mut reader).await);
-
-        let mut buf = Vec::new();
-        let outcome = read_bounded_line(&mut reader, 1024, &mut buf)
-            .await
-            .expect("next line must still be readable");
-        assert_eq!(outcome, BoundedReadOutcome::Line { body_len: 9 });
-        assert_eq!(buf, b"next-line\n");
-    }
-
-    // ---- SandboxError → ActionError conversion -----------------------
-
-    #[test]
-    fn plugin_line_too_large_converts_to_fatal_action_error() {
-        let sandbox_err = SandboxError::PluginLineTooLarge {
-            limit: 1024,
-            observed: 2048,
-        };
-        let ae = sandbox_error_to_action_error(sandbox_err);
-        // Fatal classification is the contract: we do not want the
-        // engine to quietly retry a DoS attempt on a fresh connection.
-        assert!(
-            matches!(ae, ActionError::Fatal { .. }),
-            "PluginLineTooLarge must classify as Fatal, got {ae:?}",
-        );
-    }
-
-    #[test]
-    fn handshake_line_too_large_converts_to_fatal_action_error() {
-        let sandbox_err = SandboxError::HandshakeLineTooLarge {
-            limit: 4096,
-            observed: 8192,
-        };
-        let ae = sandbox_error_to_action_error(sandbox_err);
-        assert!(matches!(ae, ActionError::Fatal { .. }));
-    }
-
-    #[test]
-    fn plugin_closed_converts_to_retryable_action_error() {
-        // Plugin-closed is benign relative to DoS — retry to respawn is safe.
-        let sandbox_err = SandboxError::PluginClosed;
-        let ae = sandbox_error_to_action_error(sandbox_err);
-        assert!(
-            matches!(ae, ActionError::Retryable { .. }),
-            "PluginClosed should classify as Retryable, got {ae:?}",
-        );
-    }
-
-    #[test]
-    fn host_malformed_envelope_converts_to_fatal_action_error() {
-        let parse_err = serde_json::from_str::<serde_json::Value>("{")
-            .expect_err("fixture must produce serde_json::Error");
-        let ae = sandbox_error_to_action_error(SandboxError::HostMalformedEnvelope(parse_err));
-        assert!(matches!(ae, ActionError::Fatal { .. }));
-    }
-
-    // ---- #260 forged-handshake regression guard ----------------------
-
-    #[test]
-    fn validate_handshake_addr_accepts_matching_pair() {
-        let line = "NEBULA-PROTO-3|unix|/tmp/nebula-plugin-host-abc/sock\n";
-        let result = validate_handshake_addr(line, "/tmp/nebula-plugin-host-abc/sock", "unix");
-        assert!(result.is_ok(), "matching addr+kind must be accepted");
-    }
-
-    #[test]
-    fn validate_handshake_addr_rejects_forged_sibling_socket() {
-        // The exact #260 scenario: a compromised plugin prints a path
-        // that belongs to a DIFFERENT plugin's socket tree. The host
-        // allocated `/tmp/nebula-plugin-host-ours/sock`, but the plugin
-        // announces `/tmp/nebula-plugin-host-other/sock`. Must fail
-        // BEFORE `dial` so the host never connects to the sibling.
-        let line = "NEBULA-PROTO-3|unix|/tmp/nebula-plugin-host-other/sock\n";
-        let err = validate_handshake_addr(line, "/tmp/nebula-plugin-host-ours/sock", "unix")
-            .expect_err("forged sibling path must be rejected");
-        match err {
-            SandboxError::HandshakeAddrMismatch { expected, got } => {
-                assert!(
-                    expected.contains("/tmp/nebula-plugin-host-ours/sock"),
-                    "expected field should contain the host-allocated addr, got {expected:?}",
-                );
-                assert!(
-                    got.contains("/tmp/nebula-plugin-host-other/sock"),
-                    "got field should contain the announced addr, got {got:?}",
-                );
-            },
-            other => panic!("expected HandshakeAddrMismatch, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn validate_handshake_addr_rejects_mismatched_kind() {
-        // Plugin tries to smuggle a pipe address on a Unix host (or
-        // vice versa). Kind mismatch is a protocol violation and must
-        // be rejected with the same error type.
-        let line = "NEBULA-PROTO-3|pipe|some-pipe-name\n";
-        let err = validate_handshake_addr(line, "/tmp/nebula/sock", "unix")
-            .expect_err("kind mismatch must be rejected");
-        assert!(matches!(err, SandboxError::HandshakeAddrMismatch { .. }));
-    }
-
-    #[test]
-    fn validate_handshake_addr_rejects_malformed_handshake() {
-        // Missing kind/addr entirely → mismatch error (with a clear
-        // "<malformed handshake>" marker on the got side).
-        let line = "NEBULA-PROTO-3\n";
-        let err = validate_handshake_addr(line, "/tmp/nebula/sock", "unix")
-            .expect_err("malformed handshake must be rejected");
-        match err {
-            SandboxError::HandshakeAddrMismatch { got, .. } => {
-                assert!(
-                    got.contains("malformed"),
-                    "got should flag the malformed handshake, was {got:?}",
-                );
-            },
-            other => panic!("expected HandshakeAddrMismatch, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn handshake_addr_mismatch_converts_to_fatal_action_error() {
-        let err = SandboxError::HandshakeAddrMismatch {
-            expected: String::from("unix|/tmp/ok"),
-            got: String::from("unix|/tmp/evil"),
-        };
-        let ae = sandbox_error_to_action_error(err);
-        assert!(
-            matches!(ae, ActionError::Fatal { .. }),
-            "HandshakeAddrMismatch must classify as Fatal (no retry on forged handshake), got {ae:?}",
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn allocate_host_socket_addr_gives_distinct_tempdirs() {
-        // Two successive allocations must produce distinct addresses and
-        // distinct tempdirs — otherwise two concurrent plugin spawns
-        // would race on the same socket path.
-        let (a, kind_a, dir_a) = allocate_host_socket_addr().expect("alloc a");
-        let (b, kind_b, dir_b) = allocate_host_socket_addr().expect("alloc b");
-        assert_eq!(kind_a, "unix");
-        assert_eq!(kind_b, "unix");
-        assert_ne!(a, b, "two allocations must produce distinct socket paths");
-        assert_ne!(
-            dir_a.as_ref().map(|d| d.path().to_path_buf()),
-            dir_b.as_ref().map(|d| d.path().to_path_buf()),
-            "two allocations must produce distinct tempdirs",
-        );
-    }
-
-    #[cfg(windows)]
-    #[test]
-    fn allocate_host_socket_addr_gives_distinct_pipe_names() {
-        let (a, kind_a, dir_a) = allocate_host_socket_addr().expect("alloc a");
-        let (b, kind_b, dir_b) = allocate_host_socket_addr().expect("alloc b");
-        assert_eq!(kind_a, "pipe");
-        assert_eq!(kind_b, "pipe");
-        assert!(dir_a.is_none(), "no tempdir on windows");
-        assert!(dir_b.is_none(), "no tempdir on windows");
-        assert_ne!(a, b, "two allocations must produce distinct pipe names");
-        assert!(
-            a.starts_with(r"\\.\pipe\LOCAL\nebula-plugin-host-"),
-            "pipe name must carry the host-plugin prefix, was {a:?}",
-        );
-    }
-
-    // ---- race_cancel_timeout (#257 regression guard) -----------------
-
-    #[tokio::test]
-    async fn race_cancel_timeout_returns_ready_when_future_completes_first() {
-        let fut = async { 42u32 };
-        let outcome = race_cancel_timeout(fut, Duration::from_secs(1), None).await;
-        assert_eq!(outcome, RaceOutcome::Ready(42));
-    }
-
-    #[tokio::test]
-    async fn race_cancel_timeout_returns_timeout_without_cancel_arm() {
-        // A future that never completes, no cancel token → must time out
-        // within roughly the configured duration.
-        let fut = std::future::pending::<()>();
-        let outcome = race_cancel_timeout(fut, Duration::from_millis(25), None).await;
-        assert_eq!(outcome, RaceOutcome::Timeout);
-    }
-
-    #[tokio::test]
-    async fn race_cancel_timeout_observes_pre_cancelled_token_promptly() {
-        // If the cancellation token is already cancelled when we enter
-        // the race, the biased select must observe it on first poll
-        // WITHOUT polling the inner future. This is the core fix for
-        // #257: a cancelled workflow does not wait for `timeout`.
-        let token = CancellationToken::new();
-        token.cancel();
-
-        let start = std::time::Instant::now();
-        // `pending` future so the only way out is the cancel arm.
-        let fut = std::future::pending::<()>();
-        let outcome = race_cancel_timeout(fut, Duration::from_secs(30), Some(&token)).await;
-        let elapsed = start.elapsed();
-
-        assert_eq!(outcome, RaceOutcome::Cancelled);
-        assert!(
-            elapsed < Duration::from_millis(200),
-            "pre-cancelled token must resolve promptly, took {elapsed:?}",
-        );
-    }
-
-    #[tokio::test]
-    async fn race_cancel_timeout_wins_when_cancel_fires_mid_flight() {
-        // Future that never completes; fire the token shortly after the
-        // race starts. The cancel arm must win, not the timeout.
-        let token = CancellationToken::new();
-        let cancel_clone = token.clone();
-        tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_millis(20)).await;
-            cancel_clone.cancel();
-        });
-
-        let fut = std::future::pending::<()>();
-        // Long enough timeout that the cancel arm is the one that wins.
-        let outcome = race_cancel_timeout(fut, Duration::from_secs(30), Some(&token)).await;
-        assert_eq!(outcome, RaceOutcome::Cancelled);
-    }
-
-    #[tokio::test]
-    async fn race_cancel_timeout_ready_wins_over_never_fired_cancel() {
-        // Token exists but is never cancelled → the inner future's
-        // Ready result wins normally.
-        let token = CancellationToken::new();
-        let fut = async { "ok" };
-        let outcome = race_cancel_timeout(fut, Duration::from_secs(1), Some(&token)).await;
-        assert_eq!(outcome, RaceOutcome::Ready("ok"));
-    }
-
-    #[tokio::test]
-    async fn race_cancel_timeout_timeout_wins_over_never_fired_cancel() {
-        // Token exists but is never cancelled → deadline still applies.
-        let token = CancellationToken::new();
-        let fut = std::future::pending::<()>();
-        let outcome = race_cancel_timeout(fut, Duration::from_millis(25), Some(&token)).await;
-        assert_eq!(outcome, RaceOutcome::Timeout);
-    }
-
-    // ---- #285 monotonic-id + id-matching regression tests ------------
-
-    #[test]
-    fn next_envelope_id_is_monotonic_and_unique() {
-        let sandbox = ProcessSandbox::new(PathBuf::from("/nonexistent"), Duration::from_secs(1));
-        // Starts at 1 (not 0) so a default-zeroed response id is
-        // visibly stale.
-        let first = sandbox.next_envelope_id();
-        let second = sandbox.next_envelope_id();
-        let third = sandbox.next_envelope_id();
-        assert_eq!(first, 1);
-        assert_eq!(second, 2);
-        assert_eq!(third, 3);
-    }
-
-    #[test]
-    fn request_id_extracts_all_id_bearing_variants() {
-        let action = HostToPlugin::ActionInvoke {
-            id: 7,
-            action_key: String::from("k"),
-            input: serde_json::json!({}),
-        };
-        let meta = HostToPlugin::MetadataRequest { id: 8 };
-        let cancel = HostToPlugin::Cancel { id: 9 };
-        let rpc_ok = HostToPlugin::RpcResponseOk {
-            id: 10,
-            result: serde_json::json!({}),
-        };
-        let rpc_err = HostToPlugin::RpcResponseError {
-            id: 11,
-            code: String::from("c"),
-            message: String::from("m"),
-        };
-        let shutdown = HostToPlugin::Shutdown;
-
-        assert_eq!(request_id(&action), Some(7));
-        assert_eq!(request_id(&meta), Some(8));
-        assert_eq!(request_id(&cancel), Some(9));
-        assert_eq!(request_id(&rpc_ok), Some(10));
-        assert_eq!(request_id(&rpc_err), Some(11));
-        assert_eq!(
-            request_id(&shutdown),
-            None,
-            "Shutdown is one-way and has no correlation id",
-        );
-    }
-
-    #[test]
-    fn response_id_extracts_all_id_bearing_variants() {
-        let ok = PluginToHost::ActionResultOk {
-            id: 42,
-            output: serde_json::json!({}),
-        };
-        let err = PluginToHost::ActionResultError {
-            id: 43,
-            code: String::from("c"),
-            message: String::from("m"),
-            retryable: false,
-        };
-        let rpc = PluginToHost::RpcCall {
-            id: 44,
-            verb: String::from("v"),
-            params: serde_json::json!({}),
-        };
-        let meta = PluginToHost::MetadataResponse {
-            id: 45,
-            protocol_version: 3,
-            manifest: nebula_metadata::PluginManifest::builder("k", "K")
-                .build()
-                .unwrap(),
-            actions: Vec::new(),
-        };
-        let log = PluginToHost::Log {
-            level: nebula_plugin_sdk::protocol::LogLevel::Info,
-            message: String::from("hi"),
-            fields: serde_json::json!({}),
-        };
-
-        assert_eq!(response_id(&ok), Some(42));
-        assert_eq!(response_id(&err), Some(43));
-        assert_eq!(response_id(&rpc), Some(44));
-        assert_eq!(response_id(&meta), Some(45));
-        assert_eq!(
-            response_id(&log),
-            None,
-            "Log is one-way and has no correlation id",
-        );
-    }
-
-    #[test]
-    fn response_id_mismatch_converts_to_fatal_action_error() {
-        let err = SandboxError::ResponseIdMismatch {
-            expected: 42,
-            got: 41,
-        };
-        let ae = sandbox_error_to_action_error(err);
-        assert!(
-            matches!(ae, ActionError::Fatal { .. }),
-            "ResponseIdMismatch must classify as Fatal, got {ae:?}",
-        );
-    }
-
-    // ---- #257 review: narrowed dispatch retry policy -----------------
-    //
-    // The `dispatch_envelope` retry must only fire for
-    // [`SandboxError::PluginClosed`]. Retrying on `Cancelled`, `Timeout`,
-    // protocol violations, or mid-round-trip transport errors could
-    // double-invoke a non-idempotent action on the plugin side after the
-    // engine already gave up on the call.
-
-    #[test]
-    fn eof_after_send_is_terminal_not_respawnable() {
-        // PluginClosed observed by recv AFTER a successful send: the
-        // plugin may have received and begun a non-idempotent action
-        // before dying. Resending on a fresh process would
-        // double-execute, so this MUST be terminal.
-        let tde = TryDispatchError::from_sandbox_error_after_send(SandboxError::PluginClosed, true);
-        assert!(
-            !tde.is_respawnable(),
-            "EOF after a successful send must not silently resend the action, got {tde:?}",
-        );
-    }
-
-    #[test]
-    fn stale_handle_before_send_is_respawnable() {
-        // PluginClosed with no bytes written for this attempt: nothing
-        // reached a running plugin, so respawn-and-retry is safe.
-        let tde =
-            TryDispatchError::from_sandbox_error_after_send(SandboxError::PluginClosed, false);
-        assert!(
-            tde.is_respawnable(),
-            "no bytes reached a running plugin — respawn must be allowed, got {tde:?}",
-        );
-    }
-
-    #[test]
-    fn plugin_line_too_large_classifies_as_terminal() {
-        // DoS / protocol-abuse signal — MUST NOT be retried. A retry
-        // would simply respawn and forward another opportunity to abuse
-        // the cap; the security dashboard would also see one warn per
-        // attempt instead of a single clean failure.
-        let tde = TryDispatchError::from_sandbox_error_after_send(
-            SandboxError::PluginLineTooLarge {
-                limit: 1024,
-                observed: 2048,
-            },
-            true,
-        );
-        assert!(
-            !tde.is_respawnable(),
-            "PluginLineTooLarge must be Terminal (no retry), got {tde:?}",
-        );
-    }
-
-    #[test]
-    fn response_id_mismatch_classifies_as_terminal() {
-        // Protocol violation: a stale response must poison the call
-        // rather than silently retrying onto a fresh connection.
-        let tde = TryDispatchError::from_sandbox_error_after_send(
-            SandboxError::ResponseIdMismatch {
-                expected: 42,
-                got: 41,
-            },
-            true,
-        );
-        assert!(
-            !tde.is_respawnable(),
-            "ResponseIdMismatch must be Terminal, got {tde:?}",
-        );
-    }
-
-    #[test]
-    fn transport_poisoned_classifies_as_terminal() {
-        let tde =
-            TryDispatchError::from_sandbox_error_after_send(SandboxError::TransportPoisoned, true);
-        assert!(
-            !tde.is_respawnable(),
-            "TransportPoisoned must be Terminal, got {tde:?}",
-        );
-    }
-
-    #[test]
-    fn handshake_line_too_large_classifies_as_terminal() {
-        let tde = TryDispatchError::from_sandbox_error_after_send(
-            SandboxError::HandshakeLineTooLarge {
-                limit: 4096,
-                observed: 8192,
-            },
-            true,
-        );
-        assert!(
-            !tde.is_respawnable(),
-            "HandshakeLineTooLarge must be Terminal, got {tde:?}",
-        );
-    }
-
-    #[test]
-    fn malformed_envelope_classifies_as_terminal() {
-        // The plugin spoke but produced a non-envelope. The outer
-        // handle is already dropped by try_dispatch; retrying would
-        // spawn fresh and blindly resend — but the reviewer's concern
-        // is that any envelope that reached the plugin may have had a
-        // side effect. Classify terminal to match the general rule.
-        let parse_err = serde_json::from_str::<serde_json::Value>("{")
-            .expect_err("fixture must produce serde_json::Error");
-        let tde = TryDispatchError::from_sandbox_error_after_send(
-            SandboxError::MalformedEnvelope(parse_err),
-            true,
-        );
-        assert!(
-            !tde.is_respawnable(),
-            "MalformedEnvelope must be Terminal, got {tde:?}",
-        );
-    }
-
-    #[test]
-    fn host_malformed_envelope_classifies_as_terminal() {
-        // Host-side serialize failure — no bytes on the wire, but no
-        // point retrying a deterministic host bug.
-        let parse_err = serde_json::from_str::<serde_json::Value>("{")
-            .expect_err("fixture must produce serde_json::Error");
-        let tde = TryDispatchError::from_sandbox_error_after_send(
-            SandboxError::HostMalformedEnvelope(parse_err),
-            true,
-        );
-        assert!(
-            !tde.is_respawnable(),
-            "HostMalformedEnvelope must be Terminal, got {tde:?}",
-        );
-    }
-
-    #[test]
-    fn terminal_and_respawnable_into_action_error_round_trip() {
-        // Both classifications must carry the underlying ActionError
-        // through unmodified — the classification only governs the
-        // dispatch-level retry decision, never the error surfaced to
-        // the engine.
-        let respawn = TryDispatchError::Respawnable(ActionError::fatal("r"));
-        assert!(matches!(
-            respawn.into_action_error(),
-            ActionError::Fatal { .. }
-        ));
-        let terminal = TryDispatchError::Terminal(ActionError::Cancelled);
-        assert!(matches!(
-            terminal.into_action_error(),
-            ActionError::Cancelled
-        ));
-    }
-}
